c makefile

+++ src/c/bitmap.h
@@ -2,8 +2,10 @@
 
 #ifndef BITMAP_H
 #define BITMAP_H
+#include "pages.h"
+#include <stddef.h>
 
-int bitmap_get(void* bm, int ii);
-void bitmap_put(void* bm, int ii, int vv);
+size_t bitmap_get(void* bm, size_t ii);
+void bitmap_put(void* bm, size_t ii, size_t vv);
 
 #endif
